+++ src/main/java/com/daitda/hubservice/hubinventory/application/dto/command/UpdateHubInventoryCommand.java
@@ -1,4 +1,4 @@
-package com.daitda.hubservice.domain.hubinventory.application.dto.command;
+package com.daitda.hubservice.hubinventory.application.dto.command;
 
 import lombok.Builder;
 import lombok.Getter;
