+++ src/main/java/com/daitda/hubservice/hubinventory/application/result/ListHubInventoryResult.java
@@ -1,6 +1,6 @@
-package com.daitda.hubservice.domain.hubinventory.application.result;
+package com.daitda.hubservice.hubinventory.application.result;
 
-import com.daitda.hubservice.domain.hubinventory.domain.entity.HubInventory;
+import com.daitda.hubservice.hubinventory.domain.entity.HubInventory;
 
 import java.util.UUID;
