+++ src/main/java/com/daitda/hubservice/hubinventory/presentation/dto/response/ListHubInventoryResponse.java
@@ -1,6 +1,6 @@
-package com.daitda.hubservice.domain.hubinventory.presentation.dto.response;
+package com.daitda.hubservice.hubinventory.presentation.dto.response;
 
-import com.daitda.hubservice.domain.hubinventory.application.result.ListHubInventoryResult;
+import com.daitda.hubservice.hubinventory.application.result.ListHubInventoryResult;
 import lombok.Builder;
 import lombok.Getter;
